databricks-app/requirements.txt
streamlit~=1.38.0
pandas~=2.2.3
httpx~=0.27.0
python-dotenv~=1.0.0